fix: strip player colors and update ExpGainEvent type
- strip ANSI color codes from player names in AdminService's PlayerBanEvent handler to clean up external API logs
- update ExpGainEvent to use float instead of long for exp amounts and fix missing trailing newline
- add comprehensive documentation for the tower-defense-turret-exp feature including design, specs, tasks, and proposal

+++ plugin/src/main/java/plugin/session/ExpGainEvent.java
@@ -5,5 +5,5 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ExpGainEvent {
     public final Session session;
-    public final long amount;
+    public final float amount;
 }
